pyproject.toml
[tool.poetry]
name = "starlette-cbge"
version = "0.1.0"
description = "Generic endpoints implementations for starlette"
authors = ["George Bogodukhov <gvbgduh@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
starlette = "^0.12.7"
ujson = "^1.35"
python-multipart = "^0.0.5"
pyyaml = "^5.1.2"
[tool.poetry.dev-dependencies]
pytest = "^5.0"
pytest-asyncio = "^0.10.0"
mypy = "^0.720.0"
black = { version = "*", allows-prereleases = true }
databases = {extras = ["sqlite"], version = "^0.2.5"}
requests = "^2.22.0"
httpx = "^0.6.8"
requests-async = "^0.6.2"
pydantic = "^0.32.1"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
